+++ ClientApp/src/pages/institution/index.tsx
@@ -1,6 +1,6 @@
 import { Container } from "@material-ui/core";
 import * as React from "react";
-import NavBarLoggedIn from "components/shared/navbar/NavBar";
+import NavBarLoggedIn from "components/shared/navbar/NavBarLoggedIn";
 import RegistrationForm from "components/institution/InstitutionForm";
 import RegistrationInfo from "components/institution/InstitutionInfo";
 import useStyles from "components/institution/Styles";
